acf: add second backup production key
* acf: add second backup production key

This enhances the ACF service to use the second backup production key
when checking new ACFs.

Tested:
Uploading ACFs still works.
Did not test with an ACF created with the second backup production key.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

* bmc-acf: minor fixes

This fixes review comments and makes minor fixes: add parenthesis
to show operator precedence, and other minor changes.

Tested: ACF upload still works

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

* Fix clang format errors

Changes to bootstrap.sh were taken from upstream.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

* Use backup2 key in read path

The previous commits which added the backup2 key did not handle the
"read" case.  This enhanced the code to use the backup2 key for the
ACF read function.

Tested:
This was not tested with a backup2 key.
Tested the D-Bus interface with a development key.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

* Fix clang format error

Tested: compiles

Change-Id: Id97234d205035aa78a7eaeac097ee8527c072019
Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>