acfshell: Adding acfshell service for targeted acf
ACFShell is a novel service within the BMC framework, designed to execute
scripts embedded within targeted ACF. The BMCShell ACF file contains
specific fields that govern the execution of shell scripts within the BMC
environment.

```
"shellscript":"xxxxxxxx"
“executionTimeoutInSec”:60
“issueDumpOnCompletion”:”yes|no”
```
shellscript is a script to be executed on the BMC.The shellscript field
is base64 encoded.

executionTimeoutInSec is the execution timeout of the shell script in
seconds.The BMC will kill the script process when this timeout is
reached.

issueDumpOnCompletion upon completion of the script the BMC will
optionally initiate a BMC dump which will include the output of the
script execution.

Signed-off-by: Abhilash Raju <abhilash.kollam@gmail.com>